move helper fns into separate file
